+++ src/test/kotlin/com/yourssu/soongpt/common/support/config/ApplicationTest.kt
@@ -1,4 +1,4 @@
-package com.yourssu.soongpt.common.support
+package com.yourssu.soongpt.common.support.config
 
 import org.junit.jupiter.api.extension.ExtendWith
 import org.springframework.boot.test.context.SpringBootTest
